# contracts/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/me": {
      "get": {
        "operationId": "MeController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponse"
                }
              }
            }
          },
          "401": {
            "description": "Sessão ausente ou inválida.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Limite de chamadas excedido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Falha interna sem divulgação de detalhes.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Validação de sessão temporariamente indisponível.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Me"
        ]
      }
    },
    "/v1/me/entitlements": {
      "get": {
        "operationId": "MeController_entitlements",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntitlementsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Sessão ausente ou inválida.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Limite de chamadas excedido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Falha interna sem divulgação de detalhes.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Validação de sessão temporariamente indisponível.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Me"
        ]
      }
    },
...
  },
  "info": {
    "title": "Estuda Aí API",
    "description": "",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "SafeErrorResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "enum": [
              "INVALID_INPUT",
              "UNAUTHENTICATED",
              "FORBIDDEN",
              "NOT_FOUND",
              "PAYLOAD_TOO_LARGE",
              "RATE_LIMITED",
              "UNAVAILABLE",
              "INTERNAL_ERROR"
            ]
          },
          "message": {
            "type": "string",
            "description": "Mensagem segura em português, sem detalhes internos."
          },
          "requestId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "code",
          "message",
          "requestId"
        ]
      },
      "ProfileResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "plan": {
            "type": "string",
            "enum": [
              "FREE"
            ]
          }
        },
        "required": [
          "id",
          "plan"
        ]
      },
      "CapabilitiesResponse": {
        "type": "object",
        "properties": {
          "steve": {
            "type": "boolean",
            "example": true
          },
          "advancedAi": {
            "type": "boolean",
            "example": false
          },
          "generatedQuestions": {
            "type": "boolean",
            "example": false
          },
          "advancedAnalytics": {
            "type": "boolean",
            "example": false
          }
        },
        "required": [
          "steve",
          "advancedAi",
          "generatedQuestions",
          "advancedAnalytics"
        ]
      },
      "LimitsResponse": {
        "type": "object",
        "properties": {
          "steveDailyMessages": {
            "type": "number",
            "example": 10,
            "minimum": 1
          }
        },
        "required": [
          "steveDailyMessages"
        ]
      },
      "EntitlementsResponse": {
        "type": "object",
        "properties": {
          "plan": {
            "type": "string",
            "enum": [
              "FREE"
            ]
          },
          "capabilities": {
            "$ref": "#/components/schemas/CapabilitiesResponse"
          },
          "limits": {
            "$ref": "#/components/schemas/LimitsResponse"
          }
        },
        "required": [
          "plan",
          "capabilities",
          "limits"
        ]
      },
...
    }
  }
}